+++ tests/integration/verify/v2/service/entity/test_challenge.py
@@ -53,8 +53,19 @@ class ChallengeTestCase(IntegrationTestCase):
                 "expiration_date": "2015-07-30T20:00:00Z",
                 "status": "pending",
                 "responded_reason": "none",
-                "details": "{\\"message\\": \\"Hi! Mr. John Doe, would you like to sign up?\\", \\"date\\":\\"2020-07-01T12:13:14Z\\", \\"fields\\": [{\\"label\\": \\"Action\\", \\"value\\": \\"Sign up in portal\\"}]}",
-                "hidden_details": "{\\"ip\\": \\"172.168.1.234\\"}",
+                "details": {
+                    "message": "Hi! Mr. John Doe, would you like to sign up?",
+                    "date": "2020-07-01T12:13:14Z",
+                    "fields": [
+                        {
+                            "label": "Action",
+                            "value": "Sign up in portal"
+                        }
+                    ]
+                },
+                "hidden_details": {
+                    "ip": "172.168.1.234"
+                },
                 "factor_type": "push",
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -99,8 +110,19 @@ class ChallengeTestCase(IntegrationTestCase):
                 "expiration_date": "2015-07-30T20:00:00Z",
                 "status": "pending",
                 "responded_reason": "none",
-                "details": "{\\"message\\": \\"Hi! Mr. John Doe, would you like to sign up?\\", \\"fields\\": [{\\"label\\": \\"Action\\", \\"value\\": \\"Sign up in portal\\"}]}",
-                "hidden_details": "{\\"ip\\": \\"172.168.1.234\\"}",
+                "details": {
+                    "message": "Hi! Mr. John Doe, would you like to sign up?",
+                    "date": "2020-07-01T12:13:14Z",
+                    "fields": [
+                        {
+                            "label": "Action",
+                            "value": "Sign up in portal"
+                        }
+                    ]
+                },
+                "hidden_details": {
+                    "ip": "172.168.1.234"
+                },
                 "factor_type": "push",
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -172,8 +194,19 @@ class ChallengeTestCase(IntegrationTestCase):
                         "expiration_date": "2015-07-30T20:00:00Z",
                         "status": "pending",
                         "responded_reason": "none",
-                        "details": "{\\"message\\": \\"Hi! Mr. John Doe, would you like to sign up?\\", \\"fields\\": [{\\"label\\": \\"Action\\", \\"value\\": \\"Sign up in portal\\"}]}",
-                        "hidden_details": "{\\"ip\\": \\"172.168.1.234\\"}",
+                        "details": {
+                            "message": "Hi! Mr. John Doe, would you like to sign up?",
+                            "date": "2020-07-01T12:13:14Z",
+                            "fields": [
+                                {
+                                    "label": "Action",
+                                    "value": "Sign up in portal"
+                                }
+                            ]
+                        },
+                        "hidden_details": {
+                            "ip": "172.168.1.234"
+                        },
                         "factor_type": "push",
                         "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
@@ -229,8 +262,19 @@ class ChallengeTestCase(IntegrationTestCase):
                 "expiration_date": "2015-07-30T20:00:00Z",
                 "status": "approved",
                 "responded_reason": "none",
-                "details": "{\\"message\\": \\"Hi! Mr. John Doe, would you like to sign up?\\", \\"fields\\": [{\\"label\\": \\"Action\\", \\"value\\": \\"Sign up in portal\\"}]}",
-                "hidden_details": "{\\"ip\\": \\"172.168.1.234\\"}",
+                "details": {
+                    "message": "Hi! Mr. John Doe, would you like to sign up?",
+                    "date": "2020-07-01T12:13:14Z",
+                    "fields": [
+                        {
+                            "label": "Action",
+                            "value": "Sign up in portal"
+                        }
+                    ]
+                },
+                "hidden_details": {
+                    "ip": "172.168.1.234"
+                },
                 "factor_type": "push",
                 "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }