+++ tests/integration/verify/v2/service/entity/test_factor.py
@@ -20,14 +20,9 @@ class FactorTestCase(IntegrationTestCase):
         with self.assertRaises(TwilioException):
             self.client.verify.v2.services("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX") \
                                  .entities("identity") \
-                                 .factors.create(binding="binding", friendly_name="friendly_name", factor_type="push", config="config", twilio_sandbox_mode="twilio_sandbox_mode")
+                                 .factors.create(friendly_name="friendly_name", factor_type="push", twilio_sandbox_mode="twilio_sandbox_mode")
 
-        values = {
-            'Binding': "binding",
-            'FriendlyName': "friendly_name",
-            'FactorType': "push",
-            'Config': "config",
-        }
+        values = {'FriendlyName': "friendly_name", 'FactorType': "push", }
 
         headers = {'Twilio-Sandbox-Mode': "twilio_sandbox_mode", }
         self.holodeck.assert_has_request(Request(
@@ -58,7 +53,7 @@ class FactorTestCase(IntegrationTestCase):
                 "factor_type": "push",
                 "config": {
                     "sdk_version": "1.0",
-                    "app_id": "com.authy.authy",
+                    "app_id": "com.example.myapp",
                     "notification_platform": "fcm",
                     "notification_token": "test_token"
                 },
@@ -69,7 +64,7 @@ class FactorTestCase(IntegrationTestCase):
 
         actual = self.client.verify.v2.services("VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX") \
                                       .entities("identity") \
-                                      .factors.create(binding="binding", friendly_name="friendly_name", factor_type="push", config="config")
+                                      .factors.create(friendly_name="friendly_name", factor_type="push")
 
         self.assertIsNotNone(actual)
 
@@ -132,7 +127,7 @@ class FactorTestCase(IntegrationTestCase):
                 "factor_type": "push",
                 "config": {
                     "sdk_version": "1.0",
-                    "app_id": "com.authy.authy",
+                    "app_id": "com.example.myapp",
                     "notification_platform": "fcm",
                     "notification_token": "test_token"
                 },
@@ -206,7 +201,7 @@ class FactorTestCase(IntegrationTestCase):
                         "factor_type": "push",
                         "config": {
                             "sdk_version": "1.0",
-                            "app_id": "com.authy.authy",
+                            "app_id": "com.example.myapp",
                             "notification_platform": "fcm",
                             "notification_token": "test_token"
                         },
@@ -264,7 +259,7 @@ class FactorTestCase(IntegrationTestCase):
                 "factor_type": "push",
                 "config": {
                     "sdk_version": "1.0",
-                    "app_id": "com.authy.authy",
+                    "app_id": "com.example.myapp",
                     "notification_platform": "fcm",
                     "notification_token": "test_token"
                 },