+++ tests/integration/verify/v2/test_service.py
@@ -47,7 +47,8 @@ class ServiceTestCase(IntegrationTestCase):
                 "custom_code_enabled": true,
                 "push": {
                     "include_date": true,
-                    "notify_service_sid": null
+                    "apn_credential_sid": "CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                    "fcm_credential_sid": null
                 },
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
@@ -99,7 +100,8 @@ class ServiceTestCase(IntegrationTestCase):
                 "custom_code_enabled": true,
                 "push": {
                     "include_date": true,
-                    "notify_service_sid": null
+                    "apn_credential_sid": null,
+                    "fcm_credential_sid": null
                 },
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",
@@ -183,7 +185,8 @@ class ServiceTestCase(IntegrationTestCase):
                         "custom_code_enabled": true,
                         "push": {
                             "include_date": true,
-                            "notify_service_sid": null
+                            "apn_credential_sid": null,
+                            "fcm_credential_sid": null
                         },
                         "date_created": "2015-07-30T20:00:00Z",
                         "date_updated": "2015-07-30T20:00:00Z",
@@ -237,7 +240,8 @@ class ServiceTestCase(IntegrationTestCase):
                 "custom_code_enabled": true,
                 "push": {
                     "include_date": true,
-                    "notify_service_sid": null
+                    "apn_credential_sid": null,
+                    "fcm_credential_sid": "CRbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
                 },
                 "date_created": "2015-07-30T20:00:00Z",
                 "date_updated": "2015-07-30T20:00:00Z",